// back-end/apps/vocechat-web/src/types/sse.ts
export interface GroupAnnouncement {
  gid: number;
  content: string;
  created_by: number;
  created_at: number;
  updated_at: number;
}